+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+* Relaxed GEOSwift dependency requirement for CocoaPods
+* Added support for MKMultiPolyline and MKMultiPolygon
+
 ## 1.0.0
 
 * Spun out of GEOSwift as part of its v5 rewrite